simrd/requirements.txt
attrs==19.3.0
cycler==0.10.0
dill==0.3.1.1
graphviz==0.14
kiwisolver==1.2.0
matplotlib==3.2.1
multiprocess==0.70.9
numpy==1.18.5
pandas==1.0.4
pathos==0.2.5
pox==0.2.7
ppft==1.6.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
scipy==1.4.1
seaborn==0.10.1
six==1.15.0
